requirements.txt
discord.py>=2.3.2
samp-client>=1.1.0
python-dotenv>=1.0.0